LPS-184264 Rename modules and classes names from admin to theme, delete unused getThemeId function store itemValue instead of themeId
