--- modules/apps/layout/layout-admin-item-selector-api/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Layout Admin Item Selector API
-Bundle-SymbolicName: com.liferay.layout.admin.item.selector.api
-Bundle-Version: 1.0.0
-Export-Package:\
-com.liferay.layout.admin.item.selector
-com.liferay.layout.admin.item.selector.criterion
