--- modules/apps/layout/layout-admin-item-selector-web/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Layout Admin Item Selector Web
-Bundle-SymbolicName: com.liferay.layout.admin.item.selector.web
-Bundle-Version: 1.0.0
-Web-ContextPath: /layout-admin-item-selector-web
