--- modules/apps/layout/layout-admin-item-selector-web/build.gradle
@@ -1,16 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.util.taglib", version: "default"
-	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.1"
-	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
-	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
-	compileOnly project(":apps:frontend-taglib:frontend-taglib-clay")
-	compileOnly project(":apps:item-selector:item-selector-api")
-	compileOnly project(":apps:item-selector:item-selector-criteria-api")
-	compileOnly project(":apps:layout:layout-admin-api")
-	compileOnly project(":apps:layout:layout-admin-item-selector-api")
-	compileOnly project(":apps:site:site-api")
-	compileOnly project(":core:petra:petra-function")
-	compileOnly project(":core:petra:petra-string")
-}
