+++ modules/apps/layout/layout-admin-web/build.gradle
@@ -39,7 +39,7 @@ dependencies {
 	compileOnly project(":apps:item-selector:item-selector-criteria-api")
 	compileOnly project(":apps:journal:journal-api")
 	compileOnly project(":apps:layout:layout-admin-api")
-	compileOnly project(":apps:layout:layout-admin-item-selector-api")
+	compileOnly project(":apps:layout:layout-theme-item-selector-api")
 	compileOnly project(":apps:layout:layout-api")
 	compileOnly project(":apps:layout:layout-content-page-editor-api")
 	compileOnly project(":apps:layout:layout-item-selector-api")
