+++ modules/apps/layout/layout-theme-item-selector-web/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
