+++ modules/apps/layout/layout-theme-item-selector-web/src/main/java/com/liferay/layout/theme/item/selector/web/internal/LayoutThemeItemSelectorCriterionHandler.java
@@ -1,0 +1,52 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.layout.theme.item.selector.web.internal;
+
+import com.liferay.item.selector.BaseItemSelectorCriterionHandler;
+import com.liferay.item.selector.ItemSelectorCriterionHandler;
+import com.liferay.layout.theme.item.selector.criterion.LayoutThemeItemSelectorCriterion;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+
+/**
+ * @author Stefan Tanasie
+ */
+@Component(service = ItemSelectorCriterionHandler.class)
+public class LayoutThemeItemSelectorCriterionHandler
+	extends BaseItemSelectorCriterionHandler<LayoutThemeItemSelectorCriterion> {
+
+	@Override
+	public Class<LayoutThemeItemSelectorCriterion>
+		getItemSelectorCriterionClass() {
+
+		return LayoutThemeItemSelectorCriterion.class;
+	}
+
+	@Activate
+	@Override
+	protected void activate(BundleContext bundleContext) {
+		super.activate(bundleContext);
+	}
+
+	@Deactivate
+	@Override
+	protected void deactivate() {
+		super.deactivate();
+	}
+
+}
